requirements.txt
pandas
numpy
plotly
streamlit
yfinance
scikit-learn
prophet
xgboost
tensorflow-cpu
langchain
openai
chromadb
faiss-cpu
python-dotenv
pydantic
